requirements.txt
streamlit
firebase-admin
google-cloud-firestore
playwright
requests
beautifulsoup4
python-dotenv
pandas
google-generativeai
urllib3
